+++ CHANGELOG.md
@@ -10,6 +10,11 @@ This document lists the changes of all recent versions since `2.0.0`.
 * Don't install dev dependencies (`ruby`) anymore with `build_deps => true`. It's only needed on
   the Puppet Master and shouldn't be deployed onto each node.
 
+### Other notable changes
+
+* ([#185](https://github.com/willdurand/puppet-nodejs/issues/185)) Allow to specify a custom source
+  for NodeJS to override `nodejs.org/dist`.
+
 ### Further changes
 
 * Deprecated `::nodejs::npm`. The feature was always out of scope and introduced several hacks